+++ apps/erp/app/modules/items/index.ts
@@ -1,4 +1,3 @@
-export type * from "./changeOrder.types";
 export * from "./items.models";
 export * from "./items.service";
 export * from "./types";
